Move ContentFilter under Query
