+++ app/classes/query/content_filter/clade.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class ContentFilter
+class Query::ContentFilter
   # Content filter to restrict observations and names to a taxonomic clade.
   class Clade < StringFilter
     def initialize
