+++ app/classes/query/content_filter/with_specimen.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class ContentFilter
+class Query::ContentFilter
   class WithSpecimen < BooleanFilter
     def initialize
       super(
