+++ app/views/controllers/search/_advanced_search_filters.html.erb
@@ -1,35 +1,36 @@
 <!-- Filters Settings of Advanced Search Form -->
 
 <%= fields_for(:content_filter) do |fcf| %>
-  <div id="advanced_search_filters">
+<div id="advanced_search_filters">
 
-    <div class="mt-3">
-      <span class="font-weight-bold"><%= :advanced_search_filters.t %></span>
-      <p><%= :advanced_search_filters_explain.t %>.<p>
-    </div>
+  <div class="mt-3">
+    <span class="font-weight-bold"><%= :advanced_search_filters.t %></span>
+    <p><%= :advanced_search_filters_explain.t %>.
+    <p>
+  </div>
 
-    <% ContentFilter.all.each do |filter| %>
-      <% models = filter.models.map { |m| m.name.underscore }.join(" ") %>
+  <% Query::ContentFilter.all.each do |filter| %>
+  <% models = filter.models.map { |m| m.name.underscore }.join(" ") %>
 
-      <%= tag.div(
+  <%= tag.div(
         class: "form-group",
         data: { role: "filter", models: models,
                 advanced_search_target: "filter" }
       ) do %>
 
-        <% if filter.type == :boolean %>
-          <%= render partial: "advanced_search_filter_radio_buttons",
+  <% if filter.type == :boolean %>
+  <%= render partial: "advanced_search_filter_radio_buttons",
                      locals: { f: fcf, filter: filter } %>
-        <% elsif filter.type == [:string] %>
-          <%= render partial: "advanced_search_filter_text_field",
+  <% elsif filter.type == [:string] %>
+  <%= render partial: "advanced_search_filter_text_field",
                      locals: { f: fcf, filter: filter } %>
-        <% else %>
-          <% raise "Unexpected filter type #{filter.type.inspect}" %>
-        <% end %>
+  <% else %>
+  <% raise "Unexpected filter type #{filter.type.inspect}" %>
+  <% end %>
 
-      <% end %>
+  <% end %>
 
-    <% end %>
+  <% end %>
 
-  </div>
+</div>
 <% end # fields_for %>
